requirements.txt
dependency-injector
fastapi~=0.103.1
uvicorn
aiohttp
httpx
pyyaml
pytest-asyncio
pytest-cov
pydantic~=2.3.0
SQLAlchemy~=2.0.20
pytest~=7.4.2
databases
typing
pymysql
mysql-connector-python
customtkinter
django
Kivy
flask
pyramid
tornado
bottle
react
bokeh